# Use join instead of views