--- a/api/src/main/java/com/cloud/offering/ServiceOffering.java
+++ b/api/src/main/java/com/cloud/offering/ServiceOffering.java
@@ -102,7 +102,7 @@ public interface ServiceOffering extends InfrastructureEntity, InternalIdentity,
 
     boolean getDefaultUse();
 
-    String getSystemVmType();
+    String getVmType();
 
     String getDeploymentPlanner();