--- a/engine/schema/src/main/java/org/apache/cloudstack/storage/datastore/db/PrimaryDataStoreDao.java
+++ b/engine/schema/src/main/java/org/apache/cloudstack/storage/datastore/db/PrimaryDataStoreDao.java
@@ -139,4 +139,6 @@ public interface PrimaryDataStoreDao extends GenericDao<StoragePoolVO, Long> {
     Pair<List<Long>, Integer> searchForIdsAndCount(Long storagePoolId, String storagePoolName, Long zoneId,
             String path, Long podId, Long clusterId, String address, ScopeType scopeType, StoragePoolStatus status,
             String keyword, Filter searchFilter);
+
+    List<StoragePoolVO> listByIds(List<Long> ids);
 }