… framework/db/src/main/java/com/cloud/utils/db/GenericSearchBuilder.java
… framework/db/src/main/java/com/cloud/utils/db/JoinBuilder.java
… framework/db/src/main/java/com/cloud/utils/db/SearchBase.java
… framework/db/src/main/java/com/cloud/utils/db/SearchCriteria.java
… framework/db/src/test/java/com/cloud/utils/db/GenericDaoBaseTest.java
… plugins/metrics/src/main/java/org/apache/cloudstack/metrics/MetricsServiceImpl.java
… server/src/main/java/com/cloud/api/ApiDBUtils.java
… server/src/main/java/com/cloud/api/query/ViewResponseHelper.java
… server/src/main/java/com/cloud/api/query/dao/AccountJoinDao.java
… server/src/main/java/com/cloud/api/query/dao/AccountJoinDaoImpl.java
… server/src/main/java/com/cloud/api/query/dao/DiskOfferingJoinDao.java
… server/src/main/java/com/cloud/api/query/dao/DiskOfferingJoinDaoImpl.java
… server/src/main/java/com/cloud/api/query/dao/DomainJoinDao.java
… server/src/main/java/com/cloud/api/query/dao/DomainJoinDaoImpl.java
… server/src/main/java/com/cloud/api/query/dao/ServiceOfferingJoinDao.java
… server/src/main/java/com/cloud/api/query/dao/ServiceOfferingJoinDaoImpl.java
… server/src/main/java/com/cloud/api/query/dao/StoragePoolJoinDao.java
… server/src/main/java/com/cloud/api/query/dao/StoragePoolJoinDaoImpl.java
… server/src/main/java/com/cloud/api/query/dao/TemplateJoinDaoImpl.java
… server/src/main/java/com/cloud/configuration/ConfigurationManagerImpl.java
… server/src/main/java/com/cloud/network/NetworkServiceImpl.java
… server/src/main/java/org/apache/cloudstack/acl/RoleManagerImpl.java
… server/src/test/java/com/cloud/api/query/QueryManagerImplTest.java
… server/src/test/java/com/cloud/network/NetworkServiceImplTest.java
… server/src/test/java/com/cloud/user/MockUsageEventDao.java