### Description

To be merged after https://github.com/apache/cloudstack/pull/8410 to ensure this doesn't break anything.

This PR contains changes similar to https://github.com/apache/cloudstack/pull/8012 which is to directly query the tables and make required joins only when there is a requirement to filter on some other field. This will help reduce the query times and thus reduce load on the database as well.

| Resource              | Count | Before | After | Difference |
| --------------------- | ----- | ------ | ----- | ---------- |
|     volumes           | 748   | 50     | 47    | 3          |
|     networks          | 191   | 140    | 131   | 9          |
|     events            | 7515  | 61     | 31    | 30         |
|     accounts          | 52    | 100    | 94    | 6          |
|     domains           | 51    | 35     | 32    | 3          |
|     hosts             | 101   | 306    | 163   | 143        |
|     storage pools     | 100   | 41     | 40    | 1          |
|     service offerings | 603   | 53     | 41    | 12         |
|     disk offerings    | 705   | 48     | 37    | 11         |
|     templates         | 603   | 155    | 109   | 46         |


<!--- Describe your changes in DETAIL - And how has behaviour functionally changed. -->

<!-- For new features, provide link to FS, dev ML discussion etc. -->
<!-- In case of bug fix, the expected and actual behaviours, steps to reproduce. -->

<!-- When "Fixes: #<id>" is specified, the issue/PR will automatically be closed when this PR gets merged -->
<!-- For addressing multiple issues/PRs, use multiple "Fixes: #<id>" -->
<!-- Fixes: # -->

<!--- ******************************************************************************* -->
<!--- NOTE: AUTOMATION USES THE DESCRIPTIONS TO SET LABELS AND PRODUCE DOCUMENTATION. -->
<!--- PLEASE PUT AN 'X' in only **ONE** box -->
<!--- ******************************************************************************* -->

### Types of changes

- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] Enhancement (improves an existing feature and functionality)
- [ ] Cleanup (Code refactoring and cleanup, that may add test cases)
- [ ] build/CI

### Feature/Enhancement Scale or Bug Severity

#### Feature/Enhancement Scale

- [x] Major
- [ ] Minor

#### Bug Severity

- [ ] BLOCKER
- [ ] Critical
- [ ] Major
- [ ] Minor
- [ ] Trivial


### Screenshots (if appropriate):


### How Has This Been Tested?

<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, and the tests you ran to -->

#### How did you try to break this feature and the system with this change?

<!-- see how your change affects other areas of the code, etc. -->


<!-- Please read the [CONTRIBUTING](https://github.com/apache/cloudstack/blob/main/CONTRIBUTING.md) document -->