@blueorangutan 2024-01-09
<b>[SF] Trillian test result (tid-8765)</b>
Environment: xenserver-71 (x2), Advanced Networking with Mgmt server 7
Total time taken: 44377 seconds
Marvin logs: https://github.com/blueorangutan/acs-prs/releases/download/trillian/pr8321-t8765-xenserver-71.zip
Smoke tests completed. 119 look OK, 2 have errors, 0 did not run
Only failed and skipped tests results shown below:


Test | Result | Time (s) | Test File
--- | --- | --- | ---
test_04_change_offering_small | `Error` | 100.44 | test_service_offerings.py
test_06_disk_offering_strictness_false | `Error` | 148.42 | test_service_offerings.py
test_change_service_offering_for_vm_with_snapshots | `Error` | 98.27 | test_vm_snapshots.py
test_01_create_vm_snapshots | `Error` | 90.24 | test_vm_snapshots.py
test_02_revert_vm_snapshots | `Failure` | 90.12 | test_vm_snapshots.py
test_03_delete_vm_snapshots | `Failure` | 0.01 | test_vm_snapshots.py