@blueorangutan 2024-01-09
<b>[SF] Trillian test result (tid-8766)</b>
Environment: vmware-67u3 (x2), Advanced Networking with Mgmt server r8
Total time taken: 49277 seconds
Marvin logs: https://github.com/blueorangutan/acs-prs/releases/download/trillian/pr8321-t8766-vmware-67u3.zip
Smoke tests completed. 115 look OK, 6 have errors, 0 did not run
Only failed and skipped tests results shown below:


Test | Result | Time (s) | Test File
--- | --- | --- | ---
test_01_snapshot_root_disk | `Error` | 0.07 | test_snapshots.py
ContextSuite context=TestSnapshotStandaloneBackup>:setup | `Error` | 213.28 | test_snapshots.py
test_01_snapshot_usage | `Error` | 0.07 | test_usage.py
test_02_balanced_drs_algorithm | `Failure` | 130.02 | test_cluster_drs.py
ContextSuite context=TestListIdsParams>:setup | `Error` | 0.00 | test_list_ids_parameter.py
test_04_change_offering_small | `Error` | 97.46 | test_service_offerings.py
test_06_disk_offering_strictness_false | `Error` | 150.06 | test_service_offerings.py
test_change_service_offering_for_vm_with_snapshots | `Error` | 144.69 | test_vm_snapshots.py
test_01_create_vm_snapshots | `Error` | 123.73 | test_vm_snapshots.py
test_02_revert_vm_snapshots | `Failure` | 90.22 | test_vm_snapshots.py
test_03_delete_vm_snapshots | `Failure` | 0.02 | test_vm_snapshots.py