@blueorangutan 2024-01-11
<b>[SF] Trillian test result (tid-8788)</b>
Environment: xenserver-71 (x2), Advanced Networking with Mgmt server 7
Total time taken: 44397 seconds
Marvin logs: https://github.com/blueorangutan/acs-prs/releases/download/trillian/pr8321-t8788-xenserver-71.zip
Smoke tests completed. 118 look OK, 3 have errors, 0 did not run
Only failed and skipped tests results shown below:


Test | Result | Time (s) | Test File
--- | --- | --- | ---
test_02_trigger_shutdown | `Failure` | 341.81 | test_safe_shutdown.py
test_04_change_offering_small | `Error` | 100.42 | test_service_offerings.py
test_06_disk_offering_strictness_false | `Error` | 158.00 | test_service_offerings.py
test_change_service_offering_for_vm_with_snapshots | `Error` | 211.21 | test_vm_snapshots.py