@blueorangutan 2024-01-12
<b>[SF] Trillian test result (tid-8816)</b>
Environment: vmware-67u3 (x2), Advanced Networking with Mgmt server r8
Total time taken: 48584 seconds
Marvin logs: https://github.com/blueorangutan/acs-prs/releases/download/trillian/pr8321-t8816-vmware-67u3.zip
Smoke tests completed. 120 look OK, 1 have errors, 0 did not run
Only failed and skipped tests results shown below:


Test | Result | Time (s) | Test File
--- | --- | --- | ---
test_01_unmanage_vm_cycle | `Error` | 115.64 | test_vm_lifecycle_unmanage_import.py