@vladimirpetrov 2024-03-18
Tested on a simulator hypervisor, with a separate MS (4 CPUs, 8 GB RAM) and DB VM (8 CPUs, 16 GB RAM), using Apache Benchmark tool (ab), measuring mean response time from 100 executions per list operation.

**Number of resources:**
Hosts: 1001
Storage pools: 100
Volumes: 25110
Events: 155118
Instances: 8370

**Results, mean time in milliseconds**

list hosts
|            | Page=20    | Page=50    | Page=100   |
| --- | --- | --- | --- |
| Unpatched  | 304        | 676        | 1052       |
| Patched    | 259        | 647        | 1246       |
| Difference | -15%       | -4%        | 18%        |

list storage pools
|            | Page=20    | Page=50    | Page=100   |
| --- | --- | --- | --- |
| Unpatched  | 204        | 487        | 972        |
| Patched    | 193        | 381        | 843        |
| Difference | -5%        | -22%       | -13%       |

list volumes
|            | Page=20    | Page=50    | Page=100   |
| --- | --- | --- | --- |
| Unpatched  | 663        | 878        | 1205       |
| Patched    | 389        | 558        | 961        |
| Difference | -41%       | -36%       | -20%       |

list volume metrics
|            | Page=20    | Page=50    | Page=100   |
| --- | --- | --- | --- |
| Unpatched  | 699        | 1057       | 1329       |
| Patched    | 374        | 732        | 1299       |
| Difference | -46%       | -31%       | -2%        |

list events
|            | Page=20    | Page=50    | Page=100   |
| --- | --- | --- | --- |
| Unpatched  | 2609       | 2750       | 2748       |
| Patched    | 396        | 522        | 457        |
| Difference | -85%       | -81%       | -83%       |

list instance metrics
|            | Page=20    | Page=50    | Page=100   |
| --- | --- | --- | --- |
| Unpatched  | 914        | 1804       | 3273       |
| Patched    | 898        | 1650       | 3255       |
| Difference | -2%        | -9%        | -1%        |